Restore sync_media() to the runtime updater
There was a circular reference, but this time I've fixed it more simply - just factored RuntimeUpdater into its own .py file.
